requirements.txt
accelerate==0.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
diffusers==0.19.3
dill==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.2
idna==3.11
imhist==0.0.4
importlib_metadata==8.7.0
invisible-watermark==0.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
numpy>=1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python-headless
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.3.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.13.3
tqdm==4.67.1
transformers==4.30.2
triton==2.1.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
